requirements.txt
algoliasearch==1.12.0
algoliasearch-django==1.2.5
appdirs==1.4.0
asn1crypto==0.22.0
cffi==1.10.0
cryptography==2.1.4
diff-match-patch==20121119
Django==1.10.5
django-admin-tools==0.8.0
django-admin-tools-stats==0.8.0
django-bower==5.2.0
django-cache-utils==0.7.2
django-cors-headers==2.0.2
django-environ==0.4.3
django-extensions==1.7.8
django-filter==1.0.1
django-import-export==0.5.1
django-jsonfield==1.0.1
django-location-field==2.0.2
django-loginas==0.3.2
django-nvd3==0.9.7
django-qsstats-magic==0.7.2
django-tinymce==2.6.0
djangorestframework==3.5.4
elasticsearch==5.3.0
enum34==1.1.6
html5lib==1.0b8
idna==2.5
ipaddress==1.0.18
Jinja2==2.9.5
jsonfield==1.0.3
Markdown==2.6.8
MarkupSafe==0.23
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pdfkit==0.6.1
Pillow==4.0.0
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==17.0.0
pyparsing==2.1.10
PyPDF2==1.26.0
python-dateutil==2.6.0
python-nvd3==0.14.2
python-openid==2.2.5
python-slugify==1.1.4
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
tablib==0.11.4
Unidecode==0.4.20
urllib3==1.20
webencodings==0.5
xhtml2pdf==0.0.6
